+++ client/src/components/keywords/Search.js
@@ -19,7 +19,7 @@ class Search extends React.Component{
     constructor(props) {
         super(props);
         this.state ={
-            keyword: ''
+            keyword: '',
         }
     }
 
@@ -68,6 +68,7 @@ class Search extends React.Component{
         }
         return post(url, JSON.stringify(data), config);
     };
+  
 
     render(){
         return(
@@ -81,7 +82,7 @@ class Search extends React.Component{
                     style={style.search}
                     onChange={this.onChange}
                     inputProps={{ 'aria-label': 'description' }}
-                />&nbsp;&nbsp;
+                />&nbsp;&nbsp;&nbsp;
                 <Button type="submit" variant="outlined" style={style.submit}>봇</Button>
             </Form>
         </>
@@ -90,4 +91,3 @@ class Search extends React.Component{
 }
 
 export default Search;
-
